Separate library modules from runnable modules
