+++ scripts/lib/basic_youtube_upload.py
