+++ scripts/lib/constants.py
